requirements.txt
agno
openai
groq
python-dotenv
fastapi
uvicorn
